backend/requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
bcrypt==5.0.0
bertopic==0.17.4
blinker==1.9.0
boto3==1.42.67
botocore==1.42.67
cachetools==7.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
coverage==7.13.4
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.96.0
dataclasses-json==0.6.7
distro==1.9.0
docker==7.1.0
ecdsa==0.19.1
Faker==40.8.0
fastapi==0.135.1
filelock==3.25.0
filetype==1.2.0
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-genai==1.65.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==25.1.0
h11==0.16.0
hdbscan==0.8.41
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huey==2.6.0
huggingface_hub==1.5.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.17
langchain-google-genai==4.2.1
langchain-openai==1.1.10
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.12
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.18.0
networkx==3.6.1
nltk==3.9.3
numba==0.64.0
numpy==2.4.2
openai==2.24.0
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
packaging==24.2
pandas==2.3.3
passlib==1.7.4
pillow==12.1.1
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
plotly==6.6.0
pluggy==1.6.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
pypdf==6.7.5
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==82.0.0
shellingham==1.5.4
simsimd==6.5.15
six==1.17.0
skops==0.13.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.11
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
uvloop==0.22.1
vaderSentiment==3.3.2
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.6
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
